Somewhat moved stuff away from main.rs
